+++ code/core/src/main/java/io/zabrek/soulbound/listeners/death/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Entity death.
+ */
+package io.zabrek.soulbound.listeners.death;