+++ code/core/src/main/java/io/zabrek/soulbound/listeners/join/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Join Listener.
+ */
+package io.zabrek.soulbound.listeners.join;